account id no jobpositions

+++ src/modules/job-positions/job-positions.service.ts
@@ -12,35 +12,36 @@ export class JobPositionService {
     private readonly jobPositionRepository: Repository<JobPosition>,
   ) {}
 
-  async create(createJobPositionDto: CreateJobPositionDto): Promise<JobPosition> {
-    const jobPosition = this.jobPositionRepository.create(createJobPositionDto);
+  async create(createJobPositionDto: CreateJobPositionDto, accountId: number): Promise<JobPosition> {
+    const jobPosition = this.jobPositionRepository.create({
+      ...createJobPositionDto,
+      account_id: accountId,
+    });
     return this.jobPositionRepository.save(jobPosition);
   }
 
-  async findAll(): Promise<JobPosition[]> {
-    return this.jobPositionRepository.find();
+  async findAll(accountId: number): Promise<JobPosition[]> {
+    return this.jobPositionRepository.find({ where: { account_id: accountId } });
   }
 
-  async findOne(uuid: string): Promise<JobPosition> {
-    const jobPosition = await this.jobPositionRepository.findOne({ where: { uuid } });
+  async findOne(uuid: string, accountId: number): Promise<JobPosition> {
+    const jobPosition = await this.jobPositionRepository.findOne({ where: { uuid, account_id: accountId } });
     return jobPosition;
   }
 
-  async update(uuid: string, updateJobPositionDto: UpdateJobPositionDto): Promise<JobPosition> {
-    const jobPosition = await this.findOne(uuid);
-
+  async update(uuid: string, updateJobPositionDto: UpdateJobPositionDto, accountId: number): Promise<JobPosition> {
+    const jobPosition = await this.findOne(uuid, accountId);
     if (!jobPosition) {
-      throw new NotFoundException(`Cargo com UUID "${uuid}" não encontrado ao tentar atualizar.`);
+      throw new NotFoundException(`Cargo com UUID "${uuid}" não encontrado para esta conta ao tentar atualizar.`);
     }
-
     this.jobPositionRepository.merge(jobPosition, updateJobPositionDto);
     return this.jobPositionRepository.save(jobPosition);
   }
 
-  async remove(uuid: string): Promise<void> {
-    const result = await this.jobPositionRepository.delete({ uuid });
+  async remove(uuid: string, accountId: number): Promise<void> {
+    const result = await this.jobPositionRepository.delete({ uuid, account_id: accountId });
     if (result.affected === 0) {
-      throw new NotFoundException(`Cargo com UUID "${uuid}" não encontrado ao tentar deletar.`);
+      throw new NotFoundException(`Cargo com UUID "${uuid}" não encontrado para esta conta ao tentar deletar.`);
     }
   }
 }
